Wrote a few enzyme tests for App and DishEntry components
