+++ package.json
@@ -7,7 +7,8 @@
     "start": "nodemon ./server/index.js",
     "build:dev": "webpack -d --watch",
     "seed": "node db/seed.js",
-    "test": "jest"
+    "test": "jest --verbose --detectOpenHandles",
+    "test:coverage": "jest --verbose --coverage"
   },
   "author": "Anna Kazannik",
   "license": "ISC",
